requirements.txt
asgiref==3.8.1
bleach==6.1.0
crispy-bootstrap5==2024.2
Django==4.2.11
django-crispy-forms==2.1
django-markdownfield==0.11.0
importlib_metadata==7.1.0
Markdown==3.6
pillow==10.3.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.11.0
webencodings==0.5.1
zipp==3.18.1
